pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.6.2",
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-groq>=0.3.2",
    "langchain-milvus>=0.2.0",
    "langchain-openai>=0.3.23",
    "lxml>=5.4.0",
    "pillow>=11.2.1",
    "poppler-utils>=0.1.0",
    "pymilvus[model]>=2.5.11",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "python-magic-bin>=0.4.14",
    "streamlit>=1.45.1",
    "tesseract>=0.1.3",
    "tiktoken>=0.9.0",
    "torch>=2.7.1",
    "transformers>=4.52.4",
    "unstructured[all-docs]>=0.17.2",
]
